+++ src/components/download/helperFunctions.js
@@ -235,7 +235,7 @@ export const strainTSV = (dispatch, filePrefix, nodes, nodeVisibilities) => {
     const numDate = getTraitFromNode(node, "num_date");
     if (numDate) {
       const traitName = "date"; // matches use in augur metadata.tsv
-      if (!headerFields.includes(traitName)) headerFields.push(traitName);
+      headerInsert(headerFields, null, traitName)
       const numDateConfidence = getTraitFromNode(node, "num_date", {confidence: true});
       if (numDateConfidence && numDateConfidence[0] !== numDateConfidence[1]) {
         tipTraitValues[node.name][traitName] = `${numericToCalendar(numDate)} (${numericToCalendar(numDateConfidence[0])} - ${numericToCalendar(numDateConfidence[1])})`;
@@ -249,14 +249,15 @@ export const strainTSV = (dispatch, filePrefix, nodes, nodeVisibilities) => {
     const nodeAttrsToIgnore = ["author", "div", "num_date", "vaccine", "accession"];
     const traits = Object.keys(node.node_attrs).filter((k) => !nodeAttrsToIgnore.includes(k));
     for (const trait of traits) {
-      if (!headerFields.includes(trait)) headerFields.push(trait);
       const value = getTraitFromNode(node, trait);
-      if (value) {
+      if (value !== undefined) {
+        headerInsert(headerFields, null, trait);
         if (typeof value === 'string') {
           tipTraitValues[node.name][trait] = value;
           const url = getUrlFromNode(node, trait);
           if (url) {
-            tipTraitValues[node.name][trait] += `: ${url}`;
+            headerInsert(headerFields, trait, urlify(trait));
+            tipTraitValues[node.name][urlify(trait)] = url;
           }
         } else if (typeof value === "number") {
           tipTraitValues[node.name][trait] = parseFloat(value).toFixed(2);
@@ -268,10 +269,11 @@ export const strainTSV = (dispatch, filePrefix, nodes, nodeVisibilities) => {
     const fullAuthorInfo = getFullAuthorInfoFromNode(node);
     if (fullAuthorInfo) {
       const traitName = "author";
-      if (!headerFields.includes(traitName)) headerFields.push(traitName);
+      headerInsert(headerFields, null, traitName);
       tipTraitValues[node.name][traitName] = fullAuthorInfo.value;
       if (isPaperURLValid(fullAuthorInfo)) {
-        tipTraitValues[node.name][traitName] += `: ${fullAuthorInfo.paper_url}`;
+        headerInsert(headerFields, traitName, urlify(traitName));
+        tipTraitValues[node.name][urlify(traitName)] = fullAuthorInfo.paper_url;
       }
     }
 
@@ -279,7 +281,7 @@ export const strainTSV = (dispatch, filePrefix, nodes, nodeVisibilities) => {
     const vaccine = getVaccineFromNode(node);
     if (vaccine && vaccine.selection_date) {
       const traitName = "vaccine_selection_date";
-      if (!headerFields.includes(traitName)) headerFields.push(traitName);
+      headerInsert(headerFields, null, traitName);
       tipTraitValues[node.name][traitName] = vaccine.selection_date;
     }
 
@@ -287,8 +289,12 @@ export const strainTSV = (dispatch, filePrefix, nodes, nodeVisibilities) => {
     const accession = getAccessionFromNode(node);
     if (accession.accession) {
       const traitName = "accession";
-      if (!headerFields.includes(traitName)) headerFields.push(traitName);
-      tipTraitValues[node.name][traitName] = accession.accession + (accession.url ? `: ${accession.url}` : '');
+      headerInsert(headerFields, null, traitName);
+      tipTraitValues[node.name][traitName] = accession.accession;
+      if (accession.url) {
+        headerInsert(headerFields, traitName, urlify(traitName));
+        tipTraitValues[node.name][urlify(traitName)] = accession.url;
+      }
     }
   }
 
@@ -331,7 +337,7 @@ export const acknowledgmentsTSV = (dispatch, filePrefix, nodes, nodeVisibilities
     for (const traitName of traitsToExport) {
       const traitValue = getTraitFromNode(node, traitName);
       if (traitValue) {
-        if (!headerFields.includes(traitName)) headerFields.push(traitName);
+        headerInsert(headerFields, null, traitName)
         tipTraitValues[node.name][traitName] = traitValue;
       }
     }
@@ -340,10 +346,11 @@ export const acknowledgmentsTSV = (dispatch, filePrefix, nodes, nodeVisibilities
     const fullAuthorInfo = getFullAuthorInfoFromNode(node);
     if (fullAuthorInfo) {
       const traitName = "author";
-      if (!headerFields.includes(traitName)) headerFields.push(traitName);
+      headerInsert(headerFields, null, traitName)
       tipTraitValues[node.name][traitName] = fullAuthorInfo.value;
       if (isPaperURLValid(fullAuthorInfo)) {
-        tipTraitValues[node.name][traitName] += `: ${fullAuthorInfo.paper_url}`;
+        headerInsert(headerFields, traitName, urlify(traitName));
+        tipTraitValues[node.name][urlify(traitName)] = fullAuthorInfo.paper_url;
       }
     }
 
@@ -354,6 +361,34 @@ export const acknowledgmentsTSV = (dispatch, filePrefix, nodes, nodeVisibilities
   write(filename, MIME.tsv, createTsvString(Object.values(tipTraitValues), headerFields));
   dispatch(infoNotification({message: `Acknowledgments exported to ${filename}`}));
 };
+
+
+/**
+ * Inserts *el2* after *el1* in the provided *arr* array (modified in-place)
+ * if *el1* is `null` then we add *el2* to the end of the array (in-place)
+ * If *el2* is already in the *arr* nothing is done
+ */
+function headerInsert(arr, el1, el2) {
+  if (arr.includes(el2)) return
+  if (el1===null) {
+    arr.push(el2);
+    return
+  }
+  const idx1 = arr.indexOf(el1);
+  if (idx1===-1) {
+    console.warn(`Element ${el1} not present in provided array`)
+    return;
+  }
+  arr.splice(idx1+1, 0, el2);
+}
+
+/**
+ * For a column *name* return the associated column name to use for URLs
+ */
+function urlify(name) {
+  return `${name}__url`;
+}
+
 
 export const exportTree = ({dispatch, filePrefix, tree, isNewick, temporal, colorings, colorBy}) => {
   try {